+++ apps/deploy-web/src/utils/urlUtils.ts
@@ -37,6 +37,7 @@ export class UrlService {
   static userApiKeys = () => "/user/api-keys";
   static userFavorites = () => `/user/settings/favorites`;
   static userProfile = (username: string) => `/profile/${username}`;
+  static usage = () => "/usage";
   static login = (returnUrl?: string) => {
     let from = "/";
     if (returnUrl) {