+++ packages/jwt/src/index.ts
@@ -1,2 +1,3 @@
 export { JwtToken } from "./jwt-token";
 export { JwtValidator, type JwtValidationResult } from "./jwt-validator";
+export { createSignArbitraryAkashWallet, type SignArbitraryAkashWallet } from "./wallet-utils";