+++ apps/api/CHANGELOG.md
@@ -1,5 +1,27 @@
 
 
+## [2.106.1](https://github.com/akash-network/console/compare/console-api/v2.106.0...console-api/v2.106.1) (2025-07-16)
+
+
+### Bug Fixes
+
+* **billing:** return empty array for discounts when no stripe id ([#1694](https://github.com/akash-network/console/issues/1694)) ([19ec3c2](https://github.com/akash-network/console/commit/19ec3c2681bcb56adf059e7d7506e2bdb103d1ee))
+
+## [2.106.0](https://github.com/akash-network/console/compare/console-api/v2.105.1...console-api/v2.106.0) (2025-07-16)
+
+
+### Features
+
+* allow filtering stripe charges list api by date created and endingBefore cursor  ([fe986bd](https://github.com/akash-network/console/commit/fe986bdd63051c9dfe3751c1c56fab0025205c1c))
+* **billing:** add usage ui ([77b5d42](https://github.com/akash-network/console/commit/77b5d42aaf4e153a6fe9f6723567520434f3d25b))
+
+## [2.105.1](https://github.com/akash-network/console/compare/console-api/v2.105.0...console-api/v2.105.1) (2025-07-15)
+
+
+### Bug Fixes
+
+* blockchain proxy cert expired, so ignoring it ([#1685](https://github.com/akash-network/console/issues/1685)) ([3a2e205](https://github.com/akash-network/console/commit/3a2e20587918b60bb836a83704129dba27df9bae))
+
 ## [2.105.0](https://github.com/akash-network/console/compare/console-api/v2.104.0...console-api/v2.105.0) (2025-07-14)