+++ apps/deploy-web/CHANGELOG.md
@@ -1,5 +1,36 @@
 
 
+## [2.96.2](https://github.com/akash-network/console/compare/console-web/v2.96.1...console-web/v2.96.2) (2025-07-16)
+
+
+### Bug Fixes
+
+* **template:** removes client side middleware from server side props ([9da8fdc](https://github.com/akash-network/console/commit/9da8fdc313203e6b3f752ced016ce684ed4a6d33))
+
+## [2.96.1](https://github.com/akash-network/console/compare/console-web/v2.96.0...console-web/v2.96.1) (2025-07-16)
+
+
+### Bug Fixes
+
+* **alert:** properly displays deployment closed alert type in a list ([b001d29](https://github.com/akash-network/console/commit/b001d296374e4391701abd8389fdeedb2d94cd07))
+
+## [2.96.0](https://github.com/akash-network/console/compare/console-web/v2.95.1...console-web/v2.96.0) (2025-07-16)
+
+
+### Features
+
+* add calendar and date-range-picker ui components ([cc96f0f](https://github.com/akash-network/console/commit/cc96f0f71ea9078f39775623144058785b745e6c))
+* **alert:** implements deployment closed alert handler ([#1683](https://github.com/akash-network/console/issues/1683)) ([3253b75](https://github.com/akash-network/console/commit/3253b75cb2d2679996ed9df15265a5fd1a6401fe))
+* allow filtering stripe charges list api by date created and endingBefore cursor  ([fe986bd](https://github.com/akash-network/console/commit/fe986bdd63051c9dfe3751c1c56fab0025205c1c))
+* **billing:** add usage ui ([77b5d42](https://github.com/akash-network/console/commit/77b5d42aaf4e153a6fe9f6723567520434f3d25b))
+
+## [2.95.1](https://github.com/akash-network/console/compare/console-web/v2.95.0...console-web/v2.95.1) (2025-07-15)
+
+
+### Bug Fixes
+
+* blockchain proxy cert expired, so ignoring it ([#1685](https://github.com/akash-network/console/issues/1685)) ([3a2e205](https://github.com/akash-network/console/commit/3a2e20587918b60bb836a83704129dba27df9bae))
+
 ## [2.95.0](https://github.com/akash-network/console/compare/console-web/v2.94.0...console-web/v2.95.0) (2025-07-15)